+++ NEWS.md
@@ -4,7 +4,7 @@ triplot 0.3.2
 * added hierarchical_importance object (instead of 
 plot_aspects_importance_grouping)
 * triplot renamed to calculate_triplot
-* triplot is ploted by calling plot on calculate_triplot results
+* triplot is plotted by calling plot on calculate_triplot results
 
 
 triplot 0.3.1